Renamee some packages and improve on project structure
