+++ src/main/java/edu/ntnu/idi/idatt/boardgame/App.java
@@ -1,6 +1,6 @@
 package edu.ntnu.idi.idatt.boardgame;
 
-import edu.ntnu.idi.idatt.boardgame.common.view.MainView;
+import edu.ntnu.idi.idatt.boardgame.ui.MainView;
 import javafx.application.Application;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
