--- src/main/java/edu/ntnu/idi/idatt/boardgame/common/action/Action.java
@@ -1,5 +1,0 @@
-package edu.ntnu.idi.idatt.boardgame.common.action;
-
-public interface Action {
-  void execute();
-}
