--- src/main/java/edu/ntnu/idi/idatt/boardgame/common/domain/board/TileObserver.java
@@ -1,5 +1,0 @@
-package edu.ntnu.idi.idatt.boardgame.common.domain.board;
-
-public interface TileObserver {
-    void onTileChanged(Tile tile);
-}
