+++ src/main/java/edu/ntnu/idi/idatt/boardgame/core/domain/dice/DiceInterface.java
@@ -1,0 +1,5 @@
+package edu.ntnu.idi.idatt.boardgame.core.domain.dice;
+
+public interface DiceInterface {
+    int roll();
+}
