+++ src/main/java/edu/ntnu/idi/idatt/boardgame/core/domain/dice/Die.java
@@ -1,0 +1,17 @@
+package edu.ntnu.idi.idatt.boardgame.core.domain.dice;
+
+import java.util.Random;
+
+public class Die {
+  private int lastRolledValue;
+  private final Random random = new Random();
+
+  public int roll() {
+    lastRolledValue = random.nextInt(6) + 1;
+    return lastRolledValue;
+  }
+
+  public int getLastRolledValue() {
+    return lastRolledValue;
+  }
+}
