+++ src/main/java/edu/ntnu/idi/idatt/boardgame/core/engine/action/Action.java
@@ -1,0 +1,5 @@
+package edu.ntnu.idi.idatt.boardgame.core.engine.action;
+
+public interface Action {
+  void execute();
+}
