+++ src/main/java/edu/ntnu/idi/idatt/boardgame/core/engine/event/GameObserver.java
@@ -1,0 +1,8 @@
+package edu.ntnu.idi.idatt.boardgame.core.engine.event;
+
+import edu.ntnu.idi.idatt.boardgame.core.domain.player.Player;
+
+public interface GameObserver {
+    void update(String message);
+    void gameFinished(Player currentPlayer);
+}
