+++ src/main/java/edu/ntnu/idi/idatt/boardgame/core/engine/event/TileObserver.java
@@ -1,0 +1,7 @@
+package edu.ntnu.idi.idatt.boardgame.core.engine.event;
+
+import edu.ntnu.idi.idatt.boardgame.core.domain.board.Tile;
+
+public interface TileObserver {
+    void onTileChanged(Tile tile);
+}
